requirements.txt
schedule==0.5.0
redis==2.10.6
requests==2.21.0
requests-oauthlib==0.7.0
feedparser==5.2.1
pytz==2016.7
beautifulsoup4==4.5.3
numpy==1.12.1
matplotlib==2.0.0
wordcloud==1.3.1
# pyHMI==0.0.3
webdavclient==1.0.8
python-dateutil>=2.7.3
urllib3>=1.24.1
